refactor(backend): extract SQLite row decoding from openbot-database
First step of splitting the 2321-line OpenBotDatabase into controllers, following
the shape of #232. A pure move of eleven free functions, no class change: every
later step would otherwise churn the same 250 lines of helpers.
